Add EntryModalPanel, scrollbar, entries-context and SearchBar
