+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "next dev -p 3000",
     "build": "next build",
-    "start": "next start",
+    "start": "next start -p 3000",
     "lint": "next lint",
     "test": "glob -c \"node --loader tsx --no-warnings --test --watch\" \"**/?(*.)+(spec|test).[jt]s?(x)\" \"**/__tests__/**/*.[jt]s?(x)\" ",
     "prisma:seed": "pnpm exec prisma db seed",
