+++ src/app/components/EntryModalGrid/EntryModalGrid.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { Entry } from '@/utils/entries'
 import EntryModal from '@/components/EntryModal'
 
@@ -7,7 +9,7 @@ type EntryModalGridProps = {
 
 export default function EntryModalGrid({ entries }: EntryModalGridProps) {
   return (
-    <div className="grid max-w-full auto-rows-[72px]  grid-cols-[repeat(auto-fit,minmax(min(250px,100%),1fr))] gap-2">
+    <div className="grid max-w-full auto-rows-[72px] grid-cols-[repeat(auto-fit,minmax(min(250px,100%),1fr))] gap-2 p-1">
       {entries.map((entry, index) => (
         <EntryModal key={`key-${index}`} entry={entry} />
       ))}
