+++ src/app/components/Header/index.tsx
@@ -1,0 +1,10 @@
+import CompendiumHeader from './CompendiumHeader'
+import RootHeader from './Header'
+
+export default function Header() {
+  return (
+    <RootHeader>
+      <CompendiumHeader />
+    </RootHeader>
+  )
+}
