+++ src/app/components/Tabs/tabs-context.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as React from 'react'
 import { TabsContextValue, TabsProviderProps } from './types'
