+++ src/app/layout.tsx
@@ -2,6 +2,10 @@ import type { Metadata } from 'next'
 import './styles/globals.css'
 import * as fonts from './styles/fonts'
 import Header from '@/components/Header'
+import { joinClassNames } from '@/utils/index'
+import { EntriesProvider } from '@/utils/entries-context'
+import { type Entry, useEntries, dynamic } from '@/utils/entries'
+export { dynamic }
 
 export const metadata: Metadata = {
   title: 'Hyrule Compendium',
@@ -14,18 +18,1006 @@ const fontsVariables = Object.values(fonts)
   .map((font) => font.variable)
   .join(' ')
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: {
   children: React.ReactNode
 }) {
+  const { loading, error, data } = await useEntries()
+
+  if (loading) return <div>Loading...</div>
+  if (error) throw new Error('Failed loading data from database')
+  const { entries } = data as { entries: Entry[] }
+
   return (
-    <html lang="en" className={fontsVariables}>
-      <body className="flex min-h-[100dvh] flex-col bg-slate-950 px-4 font-sans text-base font-normal text-yellow-50 antialiased">
-        <Header />
-        <main className="mx-auto w-full max-w-7xl py-4">{children}</main>
-        <footer className="mt-auto">I am footer</footer>
+    <html
+      lang="en"
+      className={joinClassNames(fontsVariables, 'motion-safe:scroll-smooth')}
+    >
+      <body className="flex h-[100dvh] flex-col overflow-auto bg-slate-950 px-4 font-sans text-base font-normal text-yellow-50 antialiased">
+        <EntriesProvider entries={entries}>
+          <Header />
+          <main className="mx-auto w-full max-w-7xl flex-1 overflow-hidden py-4">
+            {children}
+          </main>
+          <footer className="">github</footer>
+        </EntriesProvider>
       </body>
     </html>
   )
 }
+
+const ENTRIES = [
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7469',
+    name: 'hyrule bass',
+    description:
+      'An ordinary fish that can be found all over Hyrule. It can be eaten raw, but cooking it amplifies its healing benefits.',
+    commonLocations: ['Hyrule Field', 'West Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa17',
+      name: 'hyrule_bass',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074569/hyrule/hyrule_bass_xyjxwo.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: '',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d746a',
+    name: 'hearty bass',
+    description:
+      'This large fish lives near the shoreline. Its sizable body can store a lot of nutrients. When cooked into a dish, it will temporarily increase your maximum hearts.',
+    commonLocations: ['West Necluda', 'Akkala Highlands'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa18',
+      name: 'hearty_bass',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074574/hyrule/hearty_bass_nksnrh.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 2,
+    cookingEffect: 'extra hearts',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d746b',
+    name: 'staminoka bass',
+    description:
+      'This Hyrule bass got to be the biggest fish by never getting caught (until now). Its long life results in a cooked dish that will restore a lot of stamina.',
+    commonLocations: ['Hyrule Field', 'West Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa19',
+      name: 'staminoka_bass',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074575/hyrule/staminoka_bass_c6ef2h.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'stamina recovery',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d746c',
+    name: 'hearty salmon',
+    description:
+      'This fish makes its home in cold water, giving it extra layers of fat. It temporarily increases your maximum hearts when used in cooking.',
+    commonLocations: ['Tabantha Frontier', 'Hebra Mountains'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa1a',
+      name: 'hearty_salmon',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074567/hyrule/hearty_salmon_qgqsy3.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 4,
+    cookingEffect: 'extra hearts',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d746d',
+    name: 'chillfin trout',
+    description:
+      'This blue trout prefers cold bodies of water. Its skin contains enzymes that keep its body cool, and when cooked into a dish, it will temporarily boost your heat resistance.',
+    commonLocations: ['Tabantha Frontier', 'Hebra Mountains'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa1b',
+      name: 'chillfin_trout',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074570/hyrule/chillfin_trout_sabxy5.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'heat resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d746e',
+    name: 'sizzlefin trout',
+    description:
+      'This red trout prefers warm bodies of water. It has a special organ specifically for storing heat. When cooked into a dish, it will temporarily increase your resistance to the cold.',
+    commonLocations: ['Eldin Canyon', 'Eldin Mountains'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa1c',
+      name: 'sizzlefin_trout',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074573/hyrule/sizzlefin_trout_xlmue0.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'cold resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d746f',
+    name: 'voltfin trout',
+    description:
+      'This trout makes its home in the freshwater lakes. Its scales have an insulating compound that, when cooked into a dish, offers resistance to electricity.',
+    commonLocations: ['Tabantha Frontier', 'Hyrule Ridge'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa1d',
+      name: 'voltfin_trout',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074569/hyrule/voltfin_trout_vgh1eu.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'shock resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7470',
+    name: 'stealthfin trout',
+    description:
+      'Consuming the bioluminescent compound that makes this fish glow in the dark will increase concentration. Dishes cooked with it will suppress noise when consumed.',
+    commonLocations: ['Great Hyrule Forest', 'Eldin Mountains'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa1e',
+      name: 'stealthfin_trout',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074570/hyrule/stealthfin_trout_igrdhc.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'stealth up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7471',
+    name: 'mighty carp',
+    description:
+      'This freshwater fish lives alongside its less mighty carp ilk. A compound in its liver promotes muscle growth. Dishes cooked with it will temporarily increase your attack power.',
+    commonLocations: ['Akkala Highlands', 'Lanayru Great Spring'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa1f',
+      name: 'mighty_carp',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074572/hyrule/mighty_carp_m6rjhj.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'attack up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7472',
+    name: 'armored carp',
+    description:
+      'Calcium deposits in the scales of this ancient fish make them as hard as armor. Cooking it into a dish will fortify your bones, temporarily increasing your defense.',
+    commonLocations: ['Lanayru Great Spring', 'East Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa20',
+      name: 'armored_carp',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074568/hyrule/armored_carp_hmgtv0.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'defense up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7473',
+    name: 'sanke carp',
+    description:
+      'This wild armored carp has been bred into a prizewinning fish. Its beautifully colored markings do not occur in nature. It offers no special effects when cooked.',
+    commonLocations: ['West Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa21',
+      name: 'sanke_carp',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074573/hyrule/sanke_carp_hash0e.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: '',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7474',
+    name: 'mighty porgy',
+    description:
+      "This ocean-dwelling fish comes with one rude attitude. The compounds in its flesh elevate your competitive spirit when it's cooked into a dish, thus increasing your attack power.",
+    commonLocations: ['Necluda Sea', 'Lanayru Sea'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa22',
+      name: 'mighty_porgy',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074571/hyrule/mighty_porgy_je5rw4.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'attack up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7475',
+    name: 'armored porgy',
+    description:
+      "This porgy's body is covered in armor-hard scales. The compounds in its scales, when cooked into a dish, fortify your bones and temporarily boost your defense.",
+    commonLocations: ['Necluda Sea', 'Lanayru Sea'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa23',
+      name: 'armored_porgy',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074560/hyrule/armored_porgy_eujj90.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'defense up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7476',
+    name: 'sneaky river snail',
+    description:
+      'This large, glow-in-the-dark snail lives in fresh water. When cooked into a dish, it heightens your senses so you can move about silently.',
+    commonLocations: ['West Necluda', 'Lanayru Great Spring'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa24',
+      name: 'sneaky_river_snail',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074573/hyrule/sneaky_river_snail_te5lmf.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'stealth up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7477',
+    name: 'hearty blueshell snail',
+    description:
+      'This snail lives on sandy beaches in large numbers. Its flesh contains a high amount of stimulants, so when cooked into a dish, it temporarily increases your maximum hearts.',
+    commonLocations: ['Lanayru Sea', 'Necluda Sea'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa25',
+      name: 'hearty_blueshell_snail',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074568/hyrule/hearty_blueshell_snail_ur2dan.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'extra hearts',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7478',
+    name: 'razorclaw crab',
+    description:
+      'This crab is well known for its exceptionally sharp pincers. When cooked, the strength compound in its claws will increase your attack power.',
+    commonLocations: ['Necluda Sea', 'East Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa26',
+      name: 'razorclaw_crab',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074575/hyrule/razorclaw_crab_tjmufu.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'attack up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7479',
+    name: 'ironshell crab',
+    description:
+      "This crab's shell is particularly hard. When cooked into dish, its fat and meat bolster the body to increase your defense.",
+    commonLocations: ['Necluda Sea', 'East Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa27',
+      name: 'ironshell_crab',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074577/hyrule/ironshell_crab_nvae4j.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'defense up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d747a',
+    name: 'bright-eyed crab',
+    description:
+      "This crab appears in large numbers when it rains. Once bite of its delectable meat, and you'll forget all your exhaustion. Replenishes your stamina when cooked into a dish.",
+    commonLocations: ['Lanayru Great Spring', 'Eldin Canyon'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa28',
+      name: 'bright-eyed_crab',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074584/hyrule/bright-eyed_crab_xc1rgw.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 1,
+    cookingEffect: 'stamina recovery',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d747b',
+    name: 'fairy',
+    description:
+      "This fairy will fly from your pouch and heal all your wounds the moment you lose your last heart. It's easily mistaken for a firefly at first, but it glows in the daylight as well as night.",
+    commonLocations: ['Greater Hyrule'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa29',
+      name: 'fairy',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074575/hyrule/fairy_ag6pkj.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 5,
+    cookingEffect: '',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d747c',
+    name: 'winterwing butterfly',
+    description:
+      "The powdery scales of this butterfly's wings cool the air around it. Watching it flutter around snowflakes is a thing of beauty. Cook it with monster parts for a heat-resistant elixir.",
+    commonLocations: ['Hyrule Ridge', 'Tabantha Frontier'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa2a',
+      name: 'winterwing_butterfly',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074573/hyrule/winterwing_butterfly_axhaf6.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'heat resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d747d',
+    name: 'summerwing butterfly',
+    description:
+      'A butterfly found in the woods and plains of warm regions. Its wings absorb the warmth of the sun. Cook it with monster parts to create an elixir that makes you feel warm and fuzzy.',
+    commonLocations: ['Great Hyrule Forest', 'Eldin Mountains'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa2b',
+      name: 'summerwing_butterfly',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074581/hyrule/summerwing_butterfly_atdrjo.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'cold resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d747e',
+    name: 'thunderwing butterfly',
+    description:
+      'This rare butterfly only shows itself when it rains. The organs in its body produce an insulating compound. When made into an elixir, it offers electrical resistance.',
+    commonLocations: ['Hyrule Ridge', 'Gerudo Highlands'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa2c',
+      name: 'thunderwing_butterfly',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074575/hyrule/thunderwing_butterfly_hid1s7.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'shock resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d747f',
+    name: 'smotherwing butterfly',
+    description:
+      'A butterfly found in the woods and plains of warm regions. Its wings absorb the warmth of the sun. Cook it with monster parts to create an elixir that makes you feel warm and fuzzy.',
+    commonLocations: ['Eldin Canyon', 'Death Mountain'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa2d',
+      name: 'smotherwing_butterfly',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074575/hyrule/smotherwing_butterfly_qe7nqt.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'flame guard',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7480',
+    name: 'cold darner',
+    description:
+      'This dragonfly prefers the cool shade of trees to the warmth of the sun. Its wings disperse heat from its body, which can be cooked into a heat-resistance elixir.',
+    commonLocations: ['Tabantha Frontier', 'Hyrule Ridge'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa2e',
+      name: 'cold_darner',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074570/hyrule/cold_darner_xafmed.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'heat resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7481',
+    name: 'warm darner',
+    description:
+      'This dragonfly has a special organ that causes it to sweat profusely. Cook it with monster parts for an elixir that will raise your core temperature so you can resist the cold.',
+    commonLocations: ['Akkala Highlands', 'Hyrule Field'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa2f',
+      name: 'warm_darner',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074575/hyrule/warm_darner_z2rdhs.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'cold resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7482',
+    name: 'electric darner',
+    description:
+      'This rare dragonfly only appears in the rain. Its wings direct electricity away from its body. Cook it with monster parts for an electricity-resistance elixir.',
+    commonLocations: ['Hyrule Ridge', 'Gerudo Desert'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa30',
+      name: 'electric_darner',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074565/hyrule/electric_darner_tgtn14.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'shock resistance',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7483',
+    name: 'restless cricket',
+    description:
+      'A very energetic cricket. Cook it with monster parts to create a stamina-recovery elixir.',
+    commonLocations: ['Hyrule Field', 'East Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa31',
+      name: 'restless_cricket',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074571/hyrule/restless_cricket_qlykjs.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'stamina recovery',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7484',
+    name: 'bladed rhino beetle',
+    description:
+      "This beetle's razor-sharp horns demand that you handle it with care. Boil the horns alongside monster parts to concoct an elixir that will raise your attack power.",
+    commonLocations: ['West Necluda', 'East Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa32',
+      name: 'bladed_rhino_beetle',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074574/hyrule/bladed_rhino_beetle_ydjgvb.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'attack up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7485',
+    name: 'rugged rhino beetle',
+    description:
+      "This beetle's hard body resembles armor. When the shell is cooked with monster parts, the resulting elixir boosts your defense.",
+    commonLocations: ['Hyrule Field', 'Faron Grasslands'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa33',
+      name: 'rugged_rhino_beetle',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074576/hyrule/rugged_rhino_beetle_fcaaxe.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'defense up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7486',
+    name: 'energetic rhino beetle',
+    description:
+      'This valuable beetle can live up to ten years. When cooked with monster parts, its impressive vitality translates into an elixir that will greatly restore your stamina.',
+    commonLocations: ['Akkala Highlands', 'West Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa34',
+      name: 'energetic_rhino_beetle',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074572/hyrule/energetic_rhino_beetle_gflswf.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'stamina recovery',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7487',
+    name: 'sunset firefly',
+    description:
+      'These fireflies glow gently in the dark. When cooked with monster parts, the compound that causes it to glow results in an elixir that will allow you to move more quietly.',
+    commonLocations: ['West Necluda', 'Great Hyrule Forest'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa35',
+      name: 'sunset_firefly',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074574/hyrule/sunset_firefly_htikwr.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'stealth up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7488',
+    name: 'hot-footed frog',
+    description:
+      'A quick frog that can be found hopping around near water. Cook it with monster parts to draw out its speed-boost effect.',
+    commonLocations: ['Lanayru Great Spring', 'Hyrule Ridge'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa36',
+      name: 'hot-footed_frog',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074575/hyrule/hot-footed_frog_jnllvx.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'speed up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d7489',
+    name: 'tireless frog',
+    description:
+      'This rare frog only ventures out in the rain. When cooked with monster parts, the elixir it produces will temporarily increase your maximum stamina.',
+    commonLocations: ['Lanayru Great Spring', 'Hyrule Ridge'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa37',
+      name: 'tireless_frog',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074572/hyrule/tireless_frog_a483i4.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'extra stamina',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d748a',
+    name: 'hightail lizard',
+    description:
+      "A lizard found throughout Hyrule. It's a but slow to react at times, but it given a chance to escape, it will dart off quickly. Cook it with monster parts for a speed-boosting elixir.",
+    commonLocations: ['West Necluda', 'East Necluda'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa38',
+      name: 'hightail_lizard',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074573/hyrule/hightail_lizard_hpxg0r.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'speed up',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d748b',
+    name: 'hearty lizard',
+    description:
+      'This rare lizard lives deep in the forests. It feeds on high-nutrient foods, giving it great vitality. When used to make elixirs, they temporarily increase your maximum hearts.',
+    commonLocations: ['Gerudo Desert', 'Necluda Sea'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa39',
+      name: 'hearty_lizard',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074577/hyrule/hearty_lizard_nko5r6.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'extra hearts',
+  },
+  {
+    __typename: 'Entry',
+    id: '648148c45e499cac399d748c',
+    name: 'fireproof lizard',
+    description:
+      'This rare lizard can only be found in the Eldin region. Its scales have heat-resistant properties, so when cooked with monster parts it produces a heat-resistance elixir.',
+    commonLocations: ['Eldin Canyon', 'Death Mountain'],
+    image: {
+      __typename: 'Image',
+      id: '6481404648f035c00925fa3a',
+      name: 'fireproof_lizard',
+      src: 'https://res.cloudinary.com/jasonzou/image/upload/v1686074577/hyrule/fireproof_lizard_umwucz.png',
+      height: 280,
+      width: 280,
+    },
+    category: {
+      __typename: 'Category',
+      id: '64813f316bbbf95ec21e3c1e',
+      name: 'food',
+    },
+    drops: [],
+    edible: true,
+    attack: null,
+    defense: null,
+    heartsRecovered: 0,
+    cookingEffect: 'flame guard',
+  },
+]
