+++ src/app/styles/globals.css
@@ -1,7 +1,36 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-/* 
-body {
-  min-height: 100dvh;
-} */
+
+@layer base {
+  /* Only for desktop users */
+  @media (min-width: 640px) {
+    html {
+      --scrollbar-background: #020617;
+      --scrollbar-thumb: #1e293b;
+      --scrollbar-thumb-hover: #334155;
+
+      /* Official styles (Firefox) */
+      scrollbar-color: var(--scrollbar-thumb) var(--scrollbar-background);
+      scrollbar-width: thin;
+    }
+
+    /* For Chrome, Edge, and Safari */
+    ::-webkit-scrollbar {
+      width: 10px;
+      background-color: var(--scrollbar-background);
+    }
+    ::-webkit-scrollbar-thumb {
+      border-radius: 1000px;
+      background-color: var(--scrollbar-thumb);
+      border: 2px solid var(--scrollbar-background);
+    }
+    /*
+	  Little bonus: on non-Firefox browsers,
+	  the thumb will light up on hover!
+	  */
+    ::-webkit-scrollbar-thumb:hover {
+      background-color: var(--scrollbar-thumb-hover);
+    }
+  }
+}
